refactor(editor): extract per-area stack order and areas panels
Mechanical module split, final step: the per-area stack order panel
and the areas cluster (area list, entity groups, badges, per-area
custom sections, async entity cache incl. its module-level helpers)
move to panels/StackOrderPanel.ts and panels/AreasPanel.ts.
StrategyEditor.ts shrinks from 5,018 to 326 lines total across the
split and keeps only host state, plumbing and the render() skeleton.

Moved-line hardening per CLAUDE.md Codacy pitfalls: areaOptionsFor()/
stateFor()/Reflect get-set-deleteProperty for all variable-key object
access, GroupOptions instead of 'as Record<string, any>' casts, inner
const arrow 'has' became a function declaration.
